+++ test/test_sem.c
@@ -3258,6 +3258,24 @@ START_TEST(test_lcs2016_04)
 }
 END_TEST
 
+START_TEST(test_lcs2016_75)
+{
+   set_standard(STD_19);
+
+   input_from_file(TESTDIR "/sem/lcs2016_75.vhd");
+
+   const error_t expect[] = {
+      { 38, "conversion only allowed between closely related types" },
+      { -1, NULL }
+   };
+   expect_errors(expect);
+
+   parse_and_check(T_ENTITY, T_ARCH);
+
+   check_expected_errors();
+}
+END_TEST
+
 Suite *get_sem_tests(void)
 {
    Suite *s = suite_create("sem");
@@ -3415,6 +3433,7 @@ Suite *get_sem_tests(void)
    tcase_add_test(tc_core, test_config2);
    tcase_add_test(tc_core, test_lcs2016_02);
    tcase_add_loop_test(tc_core, test_lcs2016_04, STD_08, STD_19 + 1);
+   tcase_add_test(tc_core, test_lcs2016_75);
    suite_add_tcase(s, tc_core);
 
    return s;