+++ NEWS.md
@@ -10,6 +10,7 @@
 - Added VHDL-2019 assert API (with @Blebowski).
 - Implemented `'image`, `'value` and `to_string` for composite types in
   VHDL-2019.
+- Implemented the "closely related record types" feature from VHDL-2019.
 
 ## Version 1.10.2 - 2023-08-20
 - Fixed a crash due to an array bounds check being incorrectly optimised